slack/pyproject.toml
[project]
name = "slack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "slack-sdk>=3.34.0",
]
